requirements.txt
amqp==2.5.0
apns2==0.5.0
asgiref==3.1.2
asn1crypto==0.24.0
astroid==2.2.5
async-timeout==2.0.1
attrs==19.1.0
autobahn==19.5.1
Automat==0.7.0
billiard==3.6.0.0
certifi==2019.3.9
cffi==1.12.3
channels==2.2.0
chardet==3.0.4
Click==7.0
constantly==15.1.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
coverage==4.0.3
cryptography==2.7
daphne==2.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.1.7
django-allauth==0.39.1
django-countries==5.3.3
django-filter==2.1.0
django-model-utils==3.1.2
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-uuidfield==0.5.0
djangorestframework==3.9.2
gunicorn==19.9.0
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.18
itypes==1.1.0
Jinja2==2.10.1
jsonfield==2.0.2
kombu==4.6.0
lazy-object-proxy==1.3.1
Markdown==3.1
MarkupSafe==1.1.1
mccabe==0.6.1
ndg-httpsclient==0.5.1
oauthlib==3.0.1
openapi-codec==1.3.2
Pillow==6.0.0
psycopg2==2.7.7
pusher==2.1.3
py-vapid==1.5.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
pyfcm==1.4.5
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
PyNaCl==1.3.0
pyOpenSSL==19.0.0
python-coveralls==2.9.1
python3-openid==3.1.0
pytz==2018.9
pywebpush==1.9.4
PyYAML==5.1
redis==3.2.1
requests==2.21.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
service-identity==18.1.0
simplejson==3.16.0
six==1.12.0
Twisted==19.2.0
txaio==18.8.1
typed-ast==1.3.5
uritemplate==3.0.0
urllib3==1.24.3
vine==1.3.0
whitenoise==4.1.2
wrapt==1.11.1
zope.interface==4.6.0
